requirements.txt
# Cliente de Integración Odoo - Dependencias

# Framework web para servidor webhook
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Cliente SignalR
signalrcore==0.9.5

# Cliente HTTP
httpx==0.25.2
requests==2.31.0

# Base de datos
aiosqlite==0.19.0

# Configuración
pyyaml==6.0.1

# Logging y monitoreo
structlog==23.2.0

# Validación de datos
pydantic==2.5.0

# Utilidades de fecha/hora
python-dateutil==2.8.2

# Utilidades async
asyncio-throttle==1.0.2

# Cliente XML-RPC mejorado (opcional, para reemplazar xmlrpc.client nativo)
# xmlrpc-client==1.0.0

# Testing (desarrollo)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Formateo de código (desarrollo)
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Documentación (desarrollo)
sphinx==7.2.6
sphinx-rtd-theme==1.3.0